#674 Rename plugins.py to base_plugin.py to match class name and move _PluginLoggerAdapter to a nested class
