+++ plugins/langfuse/langfuse_plugin.py
@@ -33,7 +33,7 @@ from langchain_core.tracers.context import register_configure_hook
 # Use lazy loading of types to avoid dependency bloat for stuff most people don't need.
 from leaf_common.config.resolver_util import ResolverUtil
 
-from neuro_san_studio.interfaces.plugins import BasePlugin
+from neuro_san_studio.interfaces.base_plugin import BasePlugin
 
 
 class LangfusePlugin(BasePlugin):
