+++ plugins/llm_config_validator/llm_config_validator_plugin.py
@@ -25,7 +25,7 @@ import asyncio
 import os
 import sys
 
-from neuro_san_studio.interfaces.plugins import BasePlugin
+from neuro_san_studio.interfaces.base_plugin import BasePlugin
 from plugins.llm_config_validator.check_llm_configs import run_checks
