+++ plugins/phoenix/phoenix_plugin.py
@@ -39,7 +39,7 @@ from typing import Type
 # Use lazy loading of types to avoid dependency bloat for stuff most people don't need.
 from leaf_common.config.resolver_util import ResolverUtil
 
-from neuro_san_studio.interfaces.plugins import BasePlugin
+from neuro_san_studio.interfaces.base_plugin import BasePlugin
 
 
 class PhoenixPlugin(BasePlugin):
