+++ tests/plugins/test_phoenix_plugin.py
@@ -18,7 +18,7 @@
 
 from unittest.mock import patch
 
-from neuro_san_studio.interfaces.plugins import BasePlugin
+from neuro_san_studio.interfaces.base_plugin import BasePlugin
 from plugins.phoenix.phoenix_plugin import PhoenixPlugin
