+++ tests/plugins/test_plugin_loader.py
@@ -19,7 +19,7 @@
 import os
 import tempfile
 
-from neuro_san_studio.interfaces.plugins import BasePlugin
+from neuro_san_studio.interfaces.base_plugin import BasePlugin
 from neuro_san_studio.plugins.factory.plugin_loader import PluginLoader
